+++ src/sora.h
@@ -39,8 +39,8 @@ class Sora : public std::enable_shared_from_this<Sora>,
   Sora(UnityContext* context);
   ~Sora();
 
-  void SetOnAddTrack(std::function<void(ptrid_t)> on_add_track);
-  void SetOnRemoveTrack(std::function<void(ptrid_t)> on_remove_track);
+  void SetOnAddTrack(std::function<void(ptrid_t, std::string)> on_add_track);
+  void SetOnRemoveTrack(std::function<void(ptrid_t, std::string)> on_remove_track);
   void SetOnNotify(std::function<void(std::string)> on_notify);
   void SetOnPush(std::function<void(std::string)> on_push);
   void SetOnMessage(std::function<void(std::string, std::string)> on_message);
@@ -116,8 +116,8 @@ class Sora : public std::enable_shared_from_this<Sora>,
   std::unique_ptr<UnityRenderer> renderer_;
   rtc::scoped_refptr<webrtc::AudioTrackInterface> audio_track_;
   rtc::scoped_refptr<webrtc::VideoTrackInterface> video_track_;
-  std::function<void(ptrid_t)> on_add_track_;
-  std::function<void(ptrid_t)> on_remove_track_;
+  std::function<void(ptrid_t, std::string)> on_add_track_;
+  std::function<void(ptrid_t, std::string)> on_remove_track_;
   std::function<void(std::string)> on_notify_;
   std::function<void(std::string)> on_push_;
   std::function<void(std::string, std::string)> on_message_;