+++ CHANGES.md
@@ -13,6 +13,8 @@
 
 ## 2022.2.1
 
+- [CHANGE] OnAddTrack, OnRemoveTrack に connectionId を追加
+    - @melpon
 - [UPDATE] Sora C++ SDK を `2022.7.8` に上げる
     - @melpon